Rename stats update scripts
